+++ modules/gif/image_frames.h
@@ -25,6 +25,8 @@ public:
 	Error load(const String &p_path, int max_frames = 0);
 	Error load_gif_from_buffer(const PoolByteArray &p_data, int max_frames = 0);
 
+	Error save_gif(const String &p_path, int p_color_count = 256);
+
 	void add_frame(const Ref<Image> &p_image, float p_delay, int p_idx = -1);
 	void remove_frame(int p_idx);