+++ tests/project/goost/core/image/test_image.gd
@@ -433,6 +433,79 @@ func test_load_gif():
 	output.unlock()
 
 
+func test_save_gif_static():
+	var image_frames = ImageFrames.new()
+
+	var frame = Image.new()
+	frame.create(64, 32, false, Image.FORMAT_RGB8)
+	frame.fill(Color.yellow)
+	image_frames.add_frame(frame, 1.0)
+
+	var err = image_frames.save_gif("res://out/static.gif")
+	assert_eq(err, OK)
+
+
+func test_save_gif_animated():
+	var image_frames = ImageFrames.new()
+
+	# Different image sizes are tested here.
+
+	var frame = Image.new()
+	frame.create(90, 90, false, Image.FORMAT_RGBA8)
+	frame.fill(Color.white)
+	image_frames.add_frame(frame, 1.0)
+
+	frame = Image.new()
+	frame.create(90, 90, false, Image.FORMAT_RGBA8)
+	frame.fill(Color.red)
+	image_frames.add_frame(frame, 1.0)
+
+	frame = Image.new()
+	frame.create(60, 60, false, Image.FORMAT_RGBA8)
+	frame.fill(Color.green)
+	image_frames.add_frame(frame, 1.0)
+
+	frame = Image.new()
+	frame.create(30, 30, false, Image.FORMAT_RGBA8)
+	frame.fill(Color.blue)
+	image_frames.add_frame(frame, 1.0)
+
+	var err = image_frames.save_gif("res://out/animated.gif")
+	assert_eq(err, OK)
+
+
+func test_save_gif_animated_rotation():
+	var image_frames = ImageFrames.new()
+
+	var angle = 0.0
+	for i in 32:
+		var frame = TestUtils.image_load(SAMPLES.icon)
+		GoostImage.rotate(frame, angle, false)
+		image_frames.add_frame(frame, 0.02)
+		angle += TAU / 32.0
+
+	var err = image_frames.save_gif("res://out/animated_rotation.gif", 60)
+	assert_eq(err, OK)
+
+
+func test_save_gif_animated_rotation_custom_indexed():
+	var image_frames = ImageFrames.new()
+
+	var angle = 0.0
+	for i in 32:
+		var frame = TestUtils.image_load(SAMPLES.icon)
+		GoostImage.rotate(frame, angle, false)
+		var indexed = ImageIndexed.new()
+		indexed.create_from_data(frame.get_width(), frame.get_height(), false, frame.get_format(), frame.get_data())
+		indexed.convert(Image.FORMAT_RGBA8)
+		indexed.generate_palette(4, ImageIndexed.DITHER_RANDOM, false, true)
+		image_frames.add_frame(indexed, 0.02)
+		angle += TAU / 32.0
+
+	var err = image_frames.save_gif("res://out/animated_rotation_custom_indexed.gif")
+	assert_eq(err, OK)
+
+
 func test_bucket_fill_4_connected():
 	var input = TestUtils.image_load(SAMPLES.icon_binary)
 	input.convert(Image.FORMAT_RGBA8)